+++ src/Pokorm.AdventOfCode/Y2023/Days/Day01.cs
@@ -6,7 +6,7 @@ public class Day01 : IDay
 
     public Day01(IInputService inputService) => this.inputService = inputService;
 
-    public int Solve()
+    public long Solve()
     {
         var lines = this.inputService.GetInputLines(2023, 1);
 
@@ -54,7 +54,7 @@ public class Day01 : IDay
         return sum;
     }
 
-    public int SolveBonus()
+    public long SolveBonus()
     {
         var lines = this.inputService.GetInputLines(2023, 1);
